Partial directory reorganization to support interfaces and cleaner abstractions
