+++ ims/einstein/filesystem/ceph.py
