+++ ims/einstein/operations.py
@@ -1,18 +1,17 @@
 #!/usr/bin/python
 import base64
-import time
-
 import os
+import time
 
 import ims.common.config as config
 import ims.common.constants as constants
 import ims.exception.db_exceptions as db_exceptions
 from ims.common.log import create_logger, log, trace
 from ims.database.database import Database
-from ims.einstein.ceph import RBD
-from ims.einstein.dnsmasq import DNSMasq
-from ims.einstein.hil import HIL
+from ims.einstein.dhcpserver.dnsmasq import DNSMasq
+from ims.einstein.filesystem.ceph import RBD
 from ims.einstein.iscsi.tgt import TGT
+from ims.einstein.networkisolator.hil import HIL
 from ims.exception.exception import RegistrationFailedException, \
     FileSystemException, DBException, HaaSException, ISCSIException, \
     AuthorizationFailedException, DHCPException
