+++ tests/picasso/test_rest.py
@@ -5,7 +5,7 @@ from unittest import TestCase
 import requests
 
 import ims.common.constants as constants
-import ims.einstein.ceph as ceph
+import ims.einstein.filesystem.ceph as ceph
 from ims.einstein.operations import BMI
 
 _cfg = config.get()
